use Rust to decode bech32 addresses
